This is a simulation of Windows 10 on Khan Academy
This will be a very detailed simulation that I'll make I'm doing this for the Virtual OS contest made by Ben Blakenship currently this is a huge WIP as I'm only starting to finish the password screen but still the deadline of this contest is 31st December and I've a lot of time till then. I'm sure when I finish this this will be one of my best programs ever.